backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
boto3==1.35.0
botocore==1.35.0
pydantic==2.8.0
pydantic-settings==2.4.0
slowapi==0.1.9
httpx==0.27.0

# ML stack
torch>=2.4.1
torchvision>=0.19.1
timm>=1.0.0
pennylane==0.44.0
numpy>=1.26.4
scipy>=1.13.0
nibabel>=5.0.0
pydicom>=2.4.0
Pillow>=9.0.0
scikit-learn>=1.4.0
kyber-py>=1.2.0
cryptography>=42.0.0
huggingface_hub>=0.23.0
